Enable serde's impls for Rc
These impls are opt-in because the behavior is potentially different than
expected -- a data structure containing multiple owners of a single Rc will
serialize the data of that Rc multiple times, and deserialization will result in
multiple distinct Rc's rather than the original data structure with a single Rc.

The behavior is the same as implemented by your serialize::{de}serialize_rc_cell.