Refactor VirtualNode, convert into enum
In the DOM, we mostly deal with elements, but there are also nodes. The
relation is that an element is a node, but not the other way around.

A node can have the following node types:

- ELEMENT_NODE
- TEXT_NODE
- COMMENT_NODE
- ..and a few others.

Previously the `VirtualNode` would handle both text nodes and elements.
A text node was a `VirtualNode` with a non-empty text field, and an
element node was a `VirtualNode` with an empty text field.

This is both a bit ugly and causes confusion. It also makes it possible
to represent illegal states (e.g. a text node with children). Therefore
this commit replaces the struct with an enum that has variants matching
the node type.

Furthermore, the `DomUpdater` now uses `Node` instead of `Element`, in
order to support both element nodes and text nodes.

Refs #68