Add html macro open tag parse tag
This commit adds a test to the `html-macro`.

The test verifies that we parse opening tags correctly, such as
```
<div id="hello">
```

This test was introduced to help others understand how to test the
macro's internals.
Being able to test the more complex aspects of the macro's internals
helps with getting quicker feedback when working on the internals.