Change `html-macro`'s `checked` attribute to specify checkedness (#206)
This PR changes the meaning of `checked` as specified in the `html!` macro.

Originally, specifying `checked` would add or remove the `checked` HTML attribute. This changed the default checkedness of the input element, but may or may not actually change the rendered state of the element.

Now, specifying `checked` directly specifies the rendered state of the element. (The default checkedness is not modified, and can be set manually.)

---

## Motivation

This change makes it easier to do the more commonly-desired task of specifying the rendered state of the checkbox (likely according to application state) rather than merely the default checkedness.

This change mitigates shooting ourselves in the foot by assuming `checked` sets a rendered checkbox's checkedness.

---

Further discussion is available in the issue #205 as well as in the new book entry and tests

https://github.com/SFBdragon/percy/tree/checked_sets_checkedness/book/src/html-macro/special-attributes

https://github.com/SFBdragon/percy/blob/checked_sets_checkedness/crates/percy-dom/tests/checked_attribute.rs

This resolves #205 for the time being.